Potential mockito server fix
