+++ libs/sdk-bindings/Cargo.toml
@@ -18,14 +18,14 @@ anyhow = { workspace = true }
 breez-sdk-core = { path = "../sdk-core" }
 sdk-common = { path = "../sdk-common" }
 thiserror = { workspace = true }
-tokio = { version = "1", features = ["full"] }
+tokio = { workspace = true }
 uniffi = { version = "0.23.0", features = ["bindgen-tests", "cli"] }
 uniffi_bindgen = "0.23.0"
 uniffi_macros = "0.23.0"
 uniffi-kotlin-multiplatform = { git = "https://gitlab.com/trixnity/uniffi-kotlin-multiplatform-bindings", rev = "bf48c5fcb153856e3055025a3cbfa56fbf213188" }
 camino = "1.1.1"
 log = { workspace = true }
-once_cell = "*"
+once_cell = { workspace = true }
 flutter_rust_bridge = "=1.82.6"
 tiny-bip39 = "*"
 tonic = { version = "^0.8", features = [
