+++ libs/sdk-core/Cargo.toml
@@ -26,7 +26,7 @@ futures = "0.3.28"
 ripemd = "0.1"
 rand = "0.8"
 tiny-bip39 = "1"
-tokio = { version = "1", features = ["full"] }
+tokio = { workspace = true }
 prost = "^0.11"
 rusqlite = { workspace = true }
 rusqlite_migration = "1.0"
@@ -42,7 +42,7 @@ tonic = { version = "^0.8", features = [
 ] }
 lazy_static = "^1.4.0"
 log = { workspace = true }
-once_cell = "1"
+once_cell = { workspace = true }
 openssl = { version = "0.10", features = ["vendored"] }
 strum = { workspace = true }
 strum_macros = { workspace = true }
@@ -55,8 +55,7 @@ serde_with = "3.3.0"
 regex = { workspace = true }
 
 [dev-dependencies]
-mockito = "1"
-#regex = "1.8.1"
+mockito = { workspace = true }
 
 [build-dependencies]
 tonic-build = "^0.8"
