+++ libs/sdk-core/src/lnurl/withdraw.rs
@@ -1,0 +1,95 @@
+#[cfg(test)]
+mod tests {
+    use anyhow::Result;
+    use mockito::Mock;
+    use sdk_common::prelude::*;
+
+    use crate::lnurl::tests::MOCK_HTTP_SERVER;
+    use crate::test_utils::rand_string;
+
+    /// Mock an LNURL-withdraw endpoint that responds with an OK to a withdraw attempt
+    fn mock_lnurl_withdraw_callback(
+        withdraw_req: &LnUrlWithdrawRequestData,
+        invoice: &LNInvoice,
+        error: Option<String>,
+    ) -> Result<Mock> {
+        let callback_url = build_withdraw_callback_url(withdraw_req, invoice)?;
+        let url = reqwest::Url::parse(&callback_url)?;
+        let mockito_path: &str = &format!("{}?{}", url.path(), url.query().unwrap());
+
+        let expected_payload = r#"
+            {"status": "OK"}
+        "#
+        .replace('\n', "");
+
+        let response_body = match error {
+            None => expected_payload,
+            Some(err_reason) => {
+                ["{\"status\": \"ERROR\", \"reason\": \"", &err_reason, "\"}"].join("")
+            }
+        };
+
+        let mut server = MOCK_HTTP_SERVER.lock().unwrap();
+        Ok(server
+            .mock("GET", mockito_path)
+            .with_body(response_body)
+            .create())
+    }
+
+    fn get_test_withdraw_req_data(min_sat: u64, max_sat: u64) -> LnUrlWithdrawRequestData {
+        LnUrlWithdrawRequestData {
+            min_withdrawable: min_sat * 1000,
+            max_withdrawable: max_sat * 1000,
+            k1: rand_string(10),
+            default_description: "test description".into(),
+            callback: "http://localhost:8080/callback".into(),
+        }
+    }
+
+    #[tokio::test]
+    async fn test_lnurl_withdraw_success() -> Result<()> {
+        let invoice_str = "lnbc110n1p38q3gtpp5ypz09jrd8p993snjwnm68cph4ftwp22le34xd4r8ftspwshxhmnsdqqxqyjw5qcqpxsp5htlg8ydpywvsa7h3u4hdn77ehs4z4e844em0apjyvmqfkzqhhd2q9qgsqqqyssqszpxzxt9uuqzymr7zxcdccj5g69s8q7zzjs7sgxn9ejhnvdh6gqjcy22mss2yexunagm5r2gqczh8k24cwrqml3njskm548aruhpwssq9nvrvz";
+        let req_invoice = crate::invoice::parse_invoice(invoice_str)?;
+        let withdraw_req = get_test_withdraw_req_data(0, 100);
+
+        let _m = mock_lnurl_withdraw_callback(&withdraw_req, &req_invoice, None)?;
+
+        assert!(matches!(
+            validate_lnurl_withdraw(withdraw_req, req_invoice.clone()).await?,
+            LnUrlWithdrawResult::Ok { data: LnUrlWithdrawSuccessData { invoice } } if invoice == req_invoice
+        ));
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_lnurl_withdraw_validate_amount_failure() -> Result<()> {
+        let invoice_str = "lnbc110n1p38q3gtpp5ypz09jrd8p993snjwnm68cph4ftwp22le34xd4r8ftspwshxhmnsdqqxqyjw5qcqpxsp5htlg8ydpywvsa7h3u4hdn77ehs4z4e844em0apjyvmqfkzqhhd2q9qgsqqqyssqszpxzxt9uuqzymr7zxcdccj5g69s8q7zzjs7sgxn9ejhnvdh6gqjcy22mss2yexunagm5r2gqczh8k24cwrqml3njskm548aruhpwssq9nvrvz";
+        let invoice = crate::invoice::parse_invoice(invoice_str)?;
+        let withdraw_req = get_test_withdraw_req_data(0, 1);
+
+        // Fail validation before even calling the endpoint (no mock needed)
+        assert!(validate_lnurl_withdraw(withdraw_req, invoice)
+            .await
+            .is_err());
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_lnurl_withdraw_endpoint_failure() -> Result<()> {
+        let invoice_str = "lnbc110n1p38q3gtpp5ypz09jrd8p993snjwnm68cph4ftwp22le34xd4r8ftspwshxhmnsdqqxqyjw5qcqpxsp5htlg8ydpywvsa7h3u4hdn77ehs4z4e844em0apjyvmqfkzqhhd2q9qgsqqqyssqszpxzxt9uuqzymr7zxcdccj5g69s8q7zzjs7sgxn9ejhnvdh6gqjcy22mss2yexunagm5r2gqczh8k24cwrqml3njskm548aruhpwssq9nvrvz";
+        let invoice = crate::invoice::parse_invoice(invoice_str)?;
+        let withdraw_req = get_test_withdraw_req_data(0, 100);
+
+        // Generic error reported by endpoint
+        let _m = mock_lnurl_withdraw_callback(&withdraw_req, &invoice, Some("error".parse()?))?;
+
+        assert!(matches!(
+            validate_lnurl_withdraw(withdraw_req, invoice).await?,
+            LnUrlWithdrawResult::ErrorStatus { data: _ }
+        ));
+
+        Ok(())
+    }
+}
